refactor: modified getting rights function for user permissions

+++ app/routes/administration.users._index/utils/delete-user.server.ts
@@ -3,6 +3,12 @@ import { Prisma } from "@prisma/client"
 import { prisma } from "~/utils/db.server"
 
 export const deleteUser = async (id: string) => {
+  if (process.env.PERMANENT_USER_ID === id) {
+    throw new Response("Error: Unable to delete the user.", {
+      status: 400,
+    })
+  }
+
   try {
     await prisma.user.delete({
       where: { id },
